Refactored the main navbar into ts
