+++ components/navbars/Menu.tsx
@@ -2,31 +2,28 @@ import styled from 'styled-components';
 import Link from 'next/link';
 import { useRouter } from 'next/router';
 
-const Menu = () => {
+export const Menu = () => {
 	const currentRoute = useRouter().pathname;
 
 	return (
 		<StyledList>
 			<StyledListItem>
-				<Link href='/dashboard'>
-					<StyledAnchor
-						isActive={currentRoute === '/dashboard' ? true : false}>
+				<Link href='/dashboard' passHref>
+					<StyledAnchor isActive={currentRoute === '/dashboard'}>
 						<img src='icons/grid.svg' alt='Dashboard' />
 					</StyledAnchor>
 				</Link>
 			</StyledListItem>
 			<StyledListItem>
-				<Link href='/workspace'>
-					<StyledAnchor
-						isActive={currentRoute === '/workspace' ? true : false}>
+				<Link href='/workspace' passHref>
+					<StyledAnchor isActive={currentRoute === '/workspace'}>
 						<img src='icons/pencil.svg' alt='Pencil icon' />
 					</StyledAnchor>
 				</Link>
 			</StyledListItem>
 			<StyledListItem>
-				<Link href='/profile'>
-					<StyledAnchor
-						isActive={currentRoute === '/profile' ? true : false}>
+				<Link href='/profile' passHref>
+					<StyledAnchor isActive={currentRoute === '/profile'}>
 						<img src='/icons/user.svg' alt='user icon' />
 					</StyledAnchor>
 				</Link>
@@ -34,8 +31,6 @@ const Menu = () => {
 		</StyledList>
 	);
 };
-
-export default Menu;
 
 const StyledList = styled.div`
 	display: flex;
@@ -65,7 +60,10 @@ const StyledListItem = styled.div`
 	top: 0;
 	width: 100%;
 `;
-const StyledAnchor = styled.a(
+interface IStyledAnchorProps {
+	isActive: boolean;
+}
+const StyledAnchor = styled.a<IStyledAnchorProps>(
 	({ isActive, theme: { colors, shadows } }) => `
 	display: flex;
 	align-items: center;
